requirements.txt
crewai>=0.80.0
crewai-tools>=0.12.0
python-dotenv>=1.0.1
requests>=2.31.0

# File parsing
pypdf>=4.2.0
python-docx>=1.1.2

# Web UI
streamlit>=1.36.0

# Utilities
pydantic>=2.8.2
